% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAR.R
\name{dose_collective}
\alias{dose_collective}
\title{Sum of the variables by given groups}
\usage{
dose_collective(data, by.var = c("year", "sex"), prefix = "LAR_", cols = NULL)
}
\arguments{
\item{data}{data table}

\item{by.var}{vector of variables to group the data, NULL if no groups}

\item{prefix}{prefix of the variables to calculate sum for, e.g. "LAR_", "Dtot", ...
If NULL, parameter \code{cols} is used.}

\item{cols}{vector with names of the variables to calculate sum for, e.g. c("LAR_tot_Colon", "LAR_tot_Thyroid").
If NULL, parameter \code{prefix} is used.}
}
\value{
data set with \code{by.var} variables and sum of the variables defined by \code{prefix} or \code{cols}.
}
\description{
Calculates sum of given variables in all rows per group.
}
\examples{
dose_collective(LAR(data = calculate_dose(dose_data),
                              dose_comp = "tot",
                              organ_ls = c("Colon", "Thyroid")),
                          prefix = "LAR_")
         dose_collective(LAR(data = calculate_dose(dose_data),
                             dose_comp = "tot",
                             organ_ls = c("Colon", "Thyroid")),
                        prefix = NULL,
                        cols = c("LAR_tot_Colon", "LAR_tot_Thyroid"))
}
\seealso{
\code{\link{calculate_dose}}, \code{\link{LAR}}
}
