% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dose_data}
\alias{dose_data}
\title{Example data set for calculating the absorbed dose}
\format{
A data set in the data.table format with 2283 rows and 10 variables:
\describe{
\item{id}{ID number (unique for each person in the database)}
\item{date_birth}{date of birth}
\item{sex}{values: W, M}
\item{household}{hunter status of household, values: Hunter, Non-hunter}
\item{event_type}{values: Cancer, Emigrated, Dead}
\item{date_event}{date of cancer diagnosis, emigration or death}
\item{year}{year of the address (of county, municipality and cesium values)}
\item{county}{code for county where the person was living}
\item{municipality}{code for municipality where the person was living}
\item{cesium}{Cs-137 surface deposition at the dwelling coordinates, in kBq/m^2}
}
}
\usage{
dose_data
}
\description{
A data set with 100 fictional Swedish inhabitants.
}
\keyword{datasets}
