% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_calc.R
\name{dose_dm}
\alias{dose_dm}
\title{Data management of the original data -- data preparation}
\usage{
dose_dm(
  data,
  Chernobyl_date = "1986-04-28",
  last_follow_up_date = NULL,
  codes_county = tab_county$Code,
  codes_municipality = tab_fshield$Code,
  ...
)
}
\arguments{
\item{data}{data table with variables as in \code{\link{dose_data}}.}

\item{Chernobyl_date}{starting date for calculations of doses,  in format "\%Y-\%m-\%d".}

\item{last_follow_up_date}{date until which the dose will be calculated,
in format "\%Y-\%m-\%d",
NULL if calculated until the last possible entry.}

\item{codes_county}{vector of county codes,
to check if all county codes in the data set are in \code{codes_county},
the rows with county code not in the \code{codes_county} are deleted.}

\item{codes_municipality}{vector of municipality codes,
to check if all municipality codes in the data set are in
\code{codes_municipality},
the rows with municipality code not in the \code{codes_municipality} are deleted.}

\item{...}{}
}
\value{
updated data set.
}
\description{
Deletes the rows with empty entries,
divides the row when a person moved to the two rows (with change of address on 30 June),
creates the variables \code{start_date} and \code{stop_date}.
}
\examples{
dose_dm(data = dose_data)
dose_dm(data = dose_data, last_follow_up_date = "1991-04-28")
}
\seealso{
\code{\link{dose_new_variables}}
}
