% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_effective.R
\name{dose_effective}
\alias{dose_effective}
\title{Calculation of the effective dose}
\usage{
dose_effective(
  data,
  tb_county = absorbedDose::tab_county,
  tb_fshield = absorbedDose::tab_fshield,
  tb_organ = absorbedDose::tab_organ,
  organ_column = "ICRP_risk_organ",
  tb_wT = absorbedDose::tab_wT,
  ...
)
}
\arguments{
\item{data}{data table with all variables as in the data set returned from
\code{\link{dose_new_variables}}.}

\item{tb_county}{table with county dependent parameters,
it should have column names like \code{\link{tab_county}}.}

\item{tb_fshield}{table with municipality dependent parameter f_shield,
it should have columns Code and \code{f_shield}, like \code{\link{tab_fshield}}.}

\item{tb_organ}{coefficients per tissue (organ),
table with columns like \code{\link{tab_organ}}.}

\item{organ_column}{name of the column in \code{tb_organ} with tissue names.
Tissue names need to be the same as in \code{tb_wT}.}

\item{tb_wT}{Tissue weighting factors,
table with columns  Organ and w_T.}

\item{...}{}
}
\value{
data set with original variables and effective dose per row,
the effective dose variables have names \code{Dext_Effective}, \code{Dint_Effective},
\code{Dmilk_Effective}, \code{Dinh_Effective}, \code{Dtot_Effective}.
}
\description{
Calculates the effective dose for every row in the data set.
The function calculates dose components
(\code{\link{dose_external}},
\code{\link{dose_internal_ing}},
\code{\link{dose_internal_milk}},
\code{\link{dose_inhalation}}, \code{\link{dose_total}}) for
all tissues for the weighted sum.
}
\examples{
dose_effective(data = dose_new_variables(dose_dm(dose_data)))
}
\seealso{
\code{\link{dose_dm}},
\code{\link{dose_new_variables}},
\code{\link{dose_external}},
\code{\link{dose_internal_ing}},
\code{\link{dose_internal_milk}},
\code{\link{dose_inhalation}}
}
