% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_calc.R
\name{dose_external}
\alias{dose_external}
\title{Calculation of the external dose}
\usage{
dose_external(
  data,
  d_Cs = 1.016856,
  phi_kerma = 0.83,
  f_out = 0.2,
  tb_Fsnow = tab_county[, .(Code, F_snow)],
  tb_fshield = tab_fshield,
  tb_organ = tab_organ,
  organ_ls = tb_organ$Organ,
  ...
)
}
\arguments{
\item{data}{data table with all variables as in the data set returned from
\code{\link{dose_new_variables}}.}

\item{d_Cs}{value in (mSv/y)/(kBq/m^2).}

\item{phi_kerma}{value in mGy/mSv.}

\item{f_out}{dimensionless value.}

\item{tb_Fsnow}{F_snow values per county, table with columns
Code and F_snow.}

\item{tb_fshield}{f_shield values per municipality,
table with columns Code and f_shield.}

\item{tb_organ}{coefficients per cancer site (organ),
table with columns
Organ, k_SEQ_organ_f, a0_f, a1_f, a2_f, a3_f,
k_SEQ_organ_m, a0_m, a1_m, a2_m, a3_m.}

\item{organ_ls}{list of cancer sites (organs) to calculate the dose for,
the organs should have coefficients in \code{tb_organ}.}

\item{...}{}
}
\value{
updated data set with external dose values per row,
the external dose variables have names of the form \code{Dext_organ_name}.
}
\description{
Calculates the external dose for every row in the data set and for
all organs given in \code{organ_ls}.
}
\examples{
dose_external(data = dose_new_variables(dose_dm(dose_data)),
organ_ls = c("Thyroid", "Colon"))
}
\seealso{
\code{\link{dose_internal_ing}},
\code{\link{dose_internal_milk}}, \code{\link{dose_inhalation}}
}
