% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_calc.R
\name{dose_inhalation}
\alias{dose_inhalation}
\title{Calculation of the inhalation dose}
\usage{
dose_inhalation(
  data,
  k_th_E = 20,
  tb_E_inh = absorbedDose::tab_county[, .(Code, E_inh)],
  organ_ls = "Thyroid",
  ...
)
}
\arguments{
\item{data}{data table with all variables as in the data set returned from
\code{\link{dose_new_variables}}.}

\item{k_th_E}{value in mGy/mSv.}

\item{tb_E_inh}{E_inh values per county, table with columns
Code and E_inh.}

\item{organ_ls}{list of cancer sites (organs) to calculate the dose for,
calculates only dose for Thyroid if it is in the \code{organ_ls}.}

\item{...}{}
}
\value{
updated data set with internal dose via diary milk ingestion values
fo Thyroid per row,
the internal dose via diary milk variable has name \code{Dinh_Thyroid}.
}
\description{
Calculates the inhalation dose for every row in the data
set for Thyroid.
}
\examples{
dose_inhalation(data = dose_new_variables(dose_dm(dose_data)),
organ_ls = c("Thyroid", "Colon"))
}
\seealso{
\code{\link{dose_external}}, \code{\link{dose_internal_ing}},
\code{\link{dose_internal_milk}}
}
