% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_calc.R
\name{dose_internal_ing}
\alias{dose_internal_ing}
\title{Calculation of the internal dose from the ingestion}
\usage{
dose_internal_ing(
  data,
  S_aliment = 1,
  FR = 0.56,
  f_sex = 0.61,
  tb_Aesd = tab_county[, .(Code, A_esd)],
  tb_organ = tab_organ,
  organ_ls = tb_organ$Organ,
  ...
)
}
\arguments{
\item{data}{data table with all variables as in the data set returned from
\code{\link{dose_new_variables}}.}

\item{S_aliment}{dimensionless value.}

\item{FR}{dimensionless value.}

\item{f_sex}{coefficients for women of age>20, men and children have
\code{f_sex=1}, dimensionless.}

\item{tb_Aesd}{A_esd values per county, table with columns
Code and A_esd.}

\item{tb_organ}{coefficients per cancer site (organ),
table with columns
Organ, Cs134_f, Cs137_f, Cs134_m,  Cs137_m.}

\item{organ_ls}{list of cancer sites (organs) to calculate the dose for,
all organs should have coefficients in \code{tb_organ}.}

\item{...}{}
}
\value{
updated data set with internal dose from the ingestion
values per row,
the internal dose variables have names of the form \code{Dint_organ_name}.
}
\description{
Calculates the internal dose from the ingestion for every row in the data set
and for all organs given in \code{organ_ls}.
}
\examples{
dose_internal_ing(data = dose_new_variables(dose_dm(dose_data)),
organ_ls = c("Thyroid", "Colon"))
}
\seealso{
\code{\link{dose_external}}, \code{\link{dose_internal_milk}}
\code{\link{dose_inhalation}}
}
