% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_calc.R
\name{dose_new_variables}
\alias{dose_new_variables}
\title{Data management of the original data -- creates new variables}
\usage{
dose_new_variables(
  data,
  Chernobyl_date = as.Date("1986-04-28", format = c("\%Y-\%m-\%d")),
  ...
)
}
\arguments{
\item{data}{data table with variables as in \code{\link{dose_data}} and
with variables \code{start_date}, \code{stop_date}.}

\item{Chernobyl_date}{starting date for calculations of doses, in format "\%Y-\%m-\%d".}

\item{...}{}
}
\value{
updated data set with the new variables.
}
\description{
Adds variables that are used in calculations of the organ absorbed dose.
The added variables are \code{start_time}, \code{stop_time}, \code{age}, \code{age_middle},
\code{stop_date_50d}, \code{age_50d}, \code{weight}.
}
\examples{
dose_new_variables(dose_dm(data = dose_data))
}
\seealso{
\code{\link{dose_dm}}
}
