% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_calc.R
\name{dose_total}
\alias{dose_total}
\title{Calculation of the total absorbed dose}
\usage{
dose_total(data, organ_ls = tab_organ$Organ, ...)
}
\arguments{
\item{data}{data table with variables for doses, like
\code{Dext_organ_name}, \code{Dint_organ_name}, \code{Dmilk_Thyroid}, \code{Dinh_Thyroid}.}

\item{organ_ls}{list of cancer sites (organs) to calculate the total dose for.}

\item{...}{}
}
\value{
updated data set with total dose per row,
the total dose variables have name of the form \code{Dtot_organ_name}.
}
\description{
Calculates the sum of the doses for every row in the data.
For each organ in \code{organ_ls}, the function adds the values of all variables that have
this organ in their name.
}
\seealso{
\code{\link{dose_external}}, \code{\link{dose_internal_ing}},
\code{\link{dose_internal_milk}}, \code{\link{dose_inhalation}}
}
