% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_calc.R
\name{dose_total_per_person}
\alias{dose_total_per_person}
\title{Calculation of the total absorbed dose per person}
\usage{
dose_total_per_person(data)
}
\arguments{
\item{data}{data table with variables for the doses calculated per row, like
\code{Dext_organ_name}, \code{Dint_organ_name}, \code{Dmilk_Thyroid}, \code{Dinh_Thyroid},
\code{Dtot_organ_name}.}
}
\value{
data table with one row per id and
variables
\code{date_birth}, \code{sex}, \code{date_event}, \code{event_type}, \code{household},
\code{start_date}, \code{stop_date} (the time period for which the dose was calculated),
\code{Dext_name_organ}, \code{Dint_name_organ}, \code{Dmilk_Thyroid}, \code{Dinh_Thyroid},
\code{Dtot_name_organ} (as sum of all rows for each id).
}
\description{
Calculation of the sum of the doses for the entire follow-up period
for every id in the data. Should be used for data set returned
from \code{\link{calculate_dose}} or
after calculating some of the dose components.
}
\examples{
dose_total_per_person(calculate_dose(data = dose_data, organ_ls = c("Thyroid", "Colon")))
}
\seealso{
\code{\link{calculate_dose}}
}
