% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAR.R
\name{LAR}
\alias{LAR}
\title{Calculation of the Lifetime attributable cancer risk (LAR)}
\usage{
LAR(
  data,
  dose_comp = c("ext", "int", "milk", "inh", "tot"),
  organ_ls = absorbedDose::tab_organ[!is.na(EPA_cancer_site), EPA_cancer_site],
  tb_LAR_M = absorbedDose::tab_LAR_M,
  tb_LAR_W = absorbedDose::tab_LAR_W,
  ...
)
}
\arguments{
\item{data}{data table with variables as in the data set returned from
\code{\link{calculate_dose}}.}

\item{dose_comp}{vector of dose components to calculate LAR for}

\item{organ_ls}{vector of cancer sites (organs) to calculate LAR for,
the doses to the organs should be in \code{data}, and LAR coefficients in \code{tb_LAR_M} and \code{tb_LAR_W}}

\item{tb_LAR_M}{LAR for cancer incidence by age at exposure for males, in cases per 10 000 person-Gy.}

\item{tb_LAR_W}{LAR for cancer incidence by age at exposure for females, in cases per 10 000 person-Gy.}

\item{...}{}
}
\value{
data set with original variables and LAR per row,
the effective dose variables have names \code{LAR_comp_organ_name}.
}
\description{
Calculates LAR for every row in the data set.
}
\examples{
tab_organ_LAR <- data.table::copy(tab_organ)
tab_organ_LAR[, Organ := EPA_cancer_site]
tab_organ_LAR <- tab_organ_LAR[!is.na(Organ), ]
LAR(data = calculate_dose(dose_data, tb_organ = tab_organ_LAR),
            dose_comp = "tot",
            organ_ls = c("Colon", "Thyroid"))
}
\seealso{
\code{\link{calculate_dose}}
}
