% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAR.R
\name{LAR_coef_lin_int}
\alias{LAR_coef_lin_int}
\title{Calculation of the age dependent value of the Lifetime attributable cancer risk (LAR)}
\usage{
LAR_coef_lin_int(age, tab)
}
\arguments{
\item{age}{a numerical vector}

\item{tab, }{LAR coefficients per cancer site, table with columns age, organ names, e.g. \code{tab_LAR_M} and \code{tab_LAR_W}}
}
\value{
data set with LAR coefficients for given age
}
\description{
For given age, gives a value of the LAR coefficients.
}
\examples{
LAR_coef_lin_int(age = c(2, 8.5, 54, 63), tab = tab_LAR_M)
}
\seealso{
\code{\link{LAR}}
}
