% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tab_county}
\alias{tab_county}
\title{County dependent parameters}
\format{
A data set in data.table format with 21 rows and 6 variables:
\describe{
\item{Code}{codes of Swedish counties, values: 1--25, without 2, 11, 15, 16}
\item{County}{names of the counties}
\item{A_esd}{average deposition of Cs-137, kBq/m^2 }
\item{F_snow}{dimensionless}
\item{c_milk}{time-integrated activity concentrations in milk, Bq d/kg}
\item{E_inh}{µSv}
}
}
\source{
Tondel M, Gabrysch K, Rääf C, Isaksson M. (2025)
A model for estimating radiation doses and population cancer risk in Sweden
after the Chernobyl Nuclear Power Plant accident in 1986. Uppsala University.
doi:10.33063/diva-544175.
}
\usage{
tab_county
}
\description{
A table with values of A_{esd}, F_{snow}, c_{milk} and E_{inh} per county.
}
\keyword{datasets}
