% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tab_fshield}
\alias{tab_fshield}
\title{Shielding factor}
\format{
A data set in data.table format with 340 rows and 3 variables:
\describe{
\item{Code}{code of Swedish municipalities}
\item{Municipality}{names of the municipalities}
\item{f_shield}{dimensionless}
}
}
\source{
Tondel, M., Gabrysch, K., Rääf, C., Isaksson, M. (2023)
Estimating the organ absorbed dose in Swedish inhabitants
following the Chernobyl Nuclear Power Plant accident with the
R package absorbedDose. Uppsala University. doi:10.33063/diva-484911.
}
\usage{
tab_fshield
}
\description{
A table with values of shielding factor f_shield per municipality.
}
\keyword{datasets}
