% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tab_LAR_W}
\alias{tab_LAR_W}
\title{LAR coefficients for females}
\format{
A data set in data.table format with 12 rows and 17 variables:
\describe{
\item{Age}{age at exposure}
\item{cancer sites}{LAR for cancer incidnce of the organ by age}
}
}
\source{
EPA (2011). Radiogenic Cancer Risk Models and Projections for the U.S. Population.
EPA 402-R-11-001. Washington, DC: U.S. Environmental Protection Agency.
}
\usage{
tab_LAR_W
}
\description{
Lifetime attributable risk (LAR) for cancer incidence (cases per 10 000 person-Gy)
by age at exposure for females. The coefficients are taken from EPA (2011).
In addition, we set coefficients to 0 at age 100.
}
\keyword{datasets}
