% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tab_municip}
\alias{tab_municip}
\title{Municipality dependent parameters}
\format{
A data set in data.table format with 290 rows and 7 variables:
\describe{
\item{Code}{code of Swedish municipalities}
\item{municipality_name}{names of the municipalities}
\item{county}{codes of Swedish counties}
\item{county_name}{names of the counties}
\item{cesium}{average deposition of Cs-137, kBq/m^2}
\item{f_shield}{dimensionless, same as in \code{\link{tab_fshield}}.}
\item{hunter_perc}{percentage of hunter households}
}
}
\source{
Tondel M, Gabrysch K, Rääf C, Isaksson M. (2025)
A model for estimating radiation doses and population cancer risk in Sweden
after the Chernobyl Nuclear Power Plant accident in 1986. Uppsala University.
doi:10.33063/diva-544175.
}
\usage{
tab_municip
}
\description{
A table with values of the average deposition, shielding factor f_shield and
percentage of hunter households per municipality.
}
\keyword{datasets}
