% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tab_organ_2023}
\alias{tab_organ_2023}
\title{Cancer site dependent parameters}
\format{
A data set in data.table format with 23 rows and 15 variables:
\describe{
\item{Organ}{names of the cancer sites}
\item{k_SEQ_organ_f, k_SEQ_organ_m}{k_SEQ values for females and males, respectively}
\item{a0_f, a1_f, a2_f, a3_f}{coefficients of the polynomial for females}
\item{a0_m, a1_m, a2_m, a3_m}{coefficients of the polynomial for males}
\item{Cs134_f, Cs134_m}{\eqn{\epsilon}
values for Cs-134 for
females and males, respectively}
\item{Cs137_f, Cs137_m}{\eqn{\epsilon}
values for Cs-137 for
females and males, respectively}
}
}
\source{
Tondel, M., Gabrysch, K., Rääf, C., Isaksson, M. (2023)
Estimating the organ absorbed dose in Swedish inhabitants
following the Chernobyl Nuclear Power Plant accident with the
R package absorbedDose. Uppsala University. doi: 10.33063/diva-484911.
}
\usage{
tab_organ_2023
}
\description{
A table with values of k_SEQ, coefficients of organ dependent polynomials,
\eqn{\epsilon}_Cs-134
and
\eqn{\epsilon}_Cs_137,
per cancer site (organ) and per sex.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Replaced with \code{\link{tab_organ}}.
}
\keyword{datasets}
