% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tab_wT}
\alias{tab_wT}
\title{Tissue weighting factors}
\format{
A data set in data.table format with 17 rows and 2 variables:
\describe{
\item{Organ}{names of the tissue}
\item{w_T}{tissue weighting factor}
}
}
\source{
ICRP (2007). ICRP Publication 103: The 2007 recommendations of the
international commission on radiological protection. Ann. ICRP 37.2–4, pp. 1–332.
doi: 10.1016/j.icrp.2007.10.003.
}
\usage{
tab_wT
}
\description{
A table with tissue weighting factors.
}
\keyword{datasets}
